Move JS to seperate folder
