+++ src/js/phrase.js
