+++ src/pages/ReceiverOptions.vue
@@ -1,6 +1,6 @@
 <script setup>
 import {VCheckbox, VTextField, VCardTitle, VCardSubtitle} from "vuetify/components";
-import {store} from "../state.js";
+import {store} from "../js/state.js";
 
 import BindPhraseInput from "../components/BindPhraseInput.vue";
 import RFSelect from "../components/RFSelect.vue";
